backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
supabase
langdetect
gTTS
numpy
python-multipart